requirements.txt
py2neo==2021.2.4
google-generativeai>=0.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.1
tqdm>=4.66.0
psutil>=5.9.8
neo4j>=5.0.0